refactor: Import Ollama from package
Previously we had to copy the ollama_functions.py from the github repo and
import the class OllamaFunctions locally as there were some changes which were
not (yet) in the old package version included (an we needed them to execute the
code). With the new package update those things are resolved.